Move zaifbot app
